requirements.txt
langchain
langchain-community
chromadb
sentence-transformers
pypdf
python-docx
tiktoken
